requirements.txt
google_api_python_client==2.111.0
gspread==5.12.2
ipython==8.16.1
ipywidgets==8.1.1
oauth2client==4.1.3
pandas==2.1.4
protobuf==4.25.1
plotly==5.18.0
